requirements.txt
vncorenlp
torch
pytorch-crf~=0.7.2